requirements.txt
-i https://pypi.org/simple
annotated-types==0.7.0; python_version >= '3.8'
anyio==4.9.0; python_version >= '3.9'
black==25.1.0; python_version >= '3.9'
certifi==2025.4.26; python_version >= '3.6'
charset-normalizer==3.4.2; python_version >= '3.7'
click==8.2.1; python_version >= '3.10'
dnspython==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
fastapi==0.115.12; python_version >= '3.8'
h11==0.16.0; python_version >= '3.8'
idna==3.10; python_version >= '3.6'
mypy-extensions==1.1.0; python_version >= '3.8'
packaging==25.0; python_version >= '3.8'
pathspec==0.12.1; python_version >= '3.8'
platformdirs==4.3.8; python_version >= '3.9'
pydantic==2.11.5; python_version >= '3.9'
pydantic-core==2.33.2; python_version >= '3.9'
pymongo[srv]==3.12.0
pytelegrambotapi==4.27.0; python_version >= '3.9'
python-dotenv==1.1.0; python_version >= '3.9'
requests==2.32.3; python_version >= '3.8'
sniffio==1.3.1; python_version >= '3.7'
starlette==0.46.2; python_version >= '3.9'
typing-extensions==4.13.2; python_version >= '3.8'
typing-inspection==0.4.1; python_version >= '3.9'
urllib3==2.4.0; python_version >= '3.9'
uvicorn==0.34.2; python_version >= '3.9'
